+++ projects/slormancer-api/src/model/content/enum/skill-genre.ts
@@ -11,4 +11,5 @@ export enum SkillGenre {
     Temporal = "temporal",
     Arcanic = "arcanic",
     Obliteration = "obliteration",
+    Fast = "fast",
 }